+++ crates/noirc_abi/src/lib.rs
@@ -67,10 +67,10 @@ impl AbiType {
     }
 
     /// Returns the number of field elements required to represent the type once encoded.
-    pub fn field_count(&self) -> usize {
+    pub fn field_count(&self) -> u32 {
         match self {
             AbiType::Field(_) | AbiType::Integer { .. } => 1,
-            AbiType::Array { visibility: _, length, typ } => typ.field_count() * (*length as usize),
+            AbiType::Array { visibility: _, length, typ } => typ.field_count() * (*length as u32),
             AbiType::Struct { fields, .. } => {
                 fields.iter().fold(0, |acc, (_, field_type)| acc + field_type.field_count())
             }
